reorganized data tests
