+++ tox.ini
@@ -37,8 +37,8 @@ commands =
     pytest -rA tests/test_components/test_types.py
     pytest -rA tests/test_components/test_viz.py
     pytest -rA tests/test_data/test_data_arrays.py
-    pytest -rA tests/test_data/test_data_monitor.py
-    pytest -rA tests/test_data/test_data_sim.py
+    pytest -rA tests/test_data/test_monitor_data.py
+    pytest -rA tests/test_data/test_sim_data.py
     pytest -rA tests/test_package/test_config.py
     pytest -rA tests/test_package/test_log.py
     pytest -rA tests/test_package/test_main.py
